requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
yt-dlp>=2024.8.6
google-generativeai>=0.7.0
python-dotenv>=1.0.0
